Arduino trying to get DHT senosors to work

+++ arduino/HalMultiSensor/HalInterfaces.h
@@ -69,7 +69,7 @@ public:
 struct TemperatureData
 {
     float temperature;
-    short humidity;
+    float humidity;
 };
 class SensorTemperature : public Sensor
 {
